+++ Tests/Helpers/FakeIlpBalanceNetworkClient+Test.swift
@@ -9,7 +9,7 @@ extension FakeIlpBalanceNetworkClient {
     ///
     /// - Parameters:
     ///     - errorResponse: An Error that will be thrown by the network client
-    public static func withErrorResponse(_ errorResponse: Error) -> FakeIlpBalanceNetworkClient {
+    public static func with(errorResponse: Error) -> FakeIlpBalanceNetworkClient {
         return FakeIlpBalanceNetworkClient(getBalanceResult: .failure(errorResponse))
     }
 
@@ -17,7 +17,7 @@ extension FakeIlpBalanceNetworkClient {
     ///
     /// - Parameters:
     ///     - statusCode: The grpc status code of the RPCError returned by the network client
-    public static func withErrorResponse(_ statusCode: StatusCode) -> FakeIlpBalanceNetworkClient {
+    public static func with(statusCode: StatusCode) -> FakeIlpBalanceNetworkClient {
         let rpcError = RPCError.callError(
             CallResult(
                 success: false,