+++ XpringKit/ILP/IlpError.swift
@@ -3,23 +3,23 @@ import SwiftGRPC
 /// Errors that may occur when interacting with Xpring's ILP infrastructure
 public enum IlpError: Error {
 
-    /// The requested functionality is not yet implemented.
-    case unimplemented
+    /// The accountID that was provided does not exist.
+    case accountNotFound
+    
+    /// Something went wrong on the ILP network
+    case internalError
 
     /// The access token has an invalid format, likely starts with "Bearer ".
     case invalidAccessToken
 
-    /// The accountID that was provided does not exist.
-    case accountNotFound
+    /// The request had an invalid argument
+    case invalidArgument
 
     /// Authentication failed on the requested connector.
     case unauthenticated
 
-    /// The request had an invalid argument
-    case invalidArgument
-
-    /// Something went wrong on the ILP network
-    case internalError
+    /// The requested functionality is not yet implemented.
+    case unimplemented
 
     /// An unknown error occured
     case unknown