+++ XpringKit/Legacy/Io_Xpring_XRPLedgerAPIServiceClient+LegacyNetworkClient.swift
@@ -1,3 +1,0 @@
-/// Conforms XRPLedgerServiceClient to NetworkClient protocol.
-extension Io_Xpring_XRPLedgerAPIServiceClient: LegacyNetworkClient {
-}